+++ package.json
@@ -88,7 +88,6 @@
     "turndown": "^7.1.1",
     "twemoji": "^1.2.0",
     "typeset": "^0.2.9",
-    "uglify-js": "^2.4.16",
     "uuid": "^3.2.1",
     "vhost": "^3.0.2",
     "xml2js": "^0.5.0",
